cleanup, minor refactor of KOS-specific channel logic
