+++ RSDKv5/RSDK/Audio/Audio.cpp
@@ -56,57 +56,6 @@ uint8 AudioDeviceBase::initializedAudioChannels = false;
 uint8 AudioDeviceBase::audioState               = 0;
 uint8 AudioDeviceBase::audioFocus               = 0;
 
-#if RETRO_PLATFORM == RETRO_KALLISTIOS
-extern "C" {
-#include "dc/sound/aica_comm.h"
-extern int snd_sh4_to_aica(void *packet, uint32_t size);
-struct snd_effect;
-typedef struct snd_effect
-{
-	uint32_t locl, locr;
-	uint32_t len;
-	uint32_t rate;
-	uint32_t used;
-	uint32_t fmt;
-	uint16_t stereo;
-
-	LIST_ENTRY(snd_effect)
-	list;
-} snd_effect_t;
-
-void snd_sfx_update_ex(sfx_play_data_t *data)
-{
-	int size;
-	snd_effect_t *t = (snd_effect_t *)data->idx;
-	AICA_CMDSTR_CHANNEL(tmp, cmd, chan);
-
-	size = t->len;
-
-	if (size >= 65535)
-		size = 65534;
-
-	cmd->cmd = AICA_CMD_CHAN;
-	cmd->timestamp = 0;
-	cmd->size = AICA_CMDSTR_CHANNEL_SIZE;
-	cmd->cmd_id = data->chn;
-
-	chan->cmd = AICA_CH_CMD_UPDATE | AICA_CH_UPDATE_SET_FREQ | AICA_CH_UPDATE_SET_PAN | AICA_CH_UPDATE_SET_VOL;
-	chan->base = t->locl;
-	chan->type = t->fmt;
-	chan->length = size;
-
-	chan->loop = data->loop;
-	chan->loopstart = data->loopstart;
-	chan->loopend = data->loopend ? data->loopend : size;
-	chan->freq = data->freq > 0 ? data->freq : t->rate;
-	chan->vol = data->vol;
-	chan->pan = data->pan;
-
-	snd_sh4_to_aica(tmp, cmd->size);
-}
-}
-#endif
-
 void AudioDeviceBase::Release()
 {
     // This is missing, meaning that the garbage collector will never reclaim stb_vorbis's buffer.
@@ -118,9 +67,10 @@ void AudioDeviceBase::Release()
 
 void AudioDeviceBase::ProcessAudioMixing(void *stream, int32 length)
 {
-//    return;
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
-    for (int32 c = 0; c < CHANNEL_COUNT; ++c) {
+    // we check every channel except the last
+    // we reserve the last channel for music stream
+    for (int32 c = 0; c < CHANNEL_COUNT - 1; ++c) {
         ChannelInfo *channel = &channels[c];
 
         switch (channel->state) {
@@ -129,21 +79,30 @@ void AudioDeviceBase::ProcessAudioMixing(void *stream, int32 length)
 
             case CHANNEL_SFX: {
                 if (channel->soundID != -1) {
-                    printf("xxx Channel %d is SFX %d\n", c, channel->soundID);
+                    // get the hardware playback channel number mapped to this RSDK ChannelInfo
                     int aicaChannel = channel->aicaChannel;
-                    if ((aicaChannel != -1)){//} && (snd_get_pos(aicaChannel) == 0)) {
-                        sfxhnd_t handle = sfxList[channel->soundID].handle;
-                        printf("xxx\ton AICA channel %d (pos %d)\n", aicaChannel, snd_get_pos(aicaChannel));
-                        double playTime = (double)(timer_ns_gettime64() - channel->startNs) * 1e-9;
+                    // valid channels are 0 throug 63
+                    if (aicaChannel != -1) {
+                        // I was hoping to use KOS APIs for the following but it isn't possible
+                        // I will explain why:
+                        // 1) `snd_is_playing` - oh, this should work just fine for this task!
+                        //    but no. it always returns true once you've fired a channel ever
+                        // 2) `snd_get_pos` - ok, I'll just check if this is at the end of the sound effect
+                        //    but no. it returns 0 if you've hit the end of the sound. you'd think, 
+                        //    ok I can check that then. also no, because some are so short you never see non-zero.
+                        // so I do this the "hard" way. when a sfx is started, I record the time since boot in NS.
+                        // every call to this function, I see if the sound should be finished playing according to:
+                        // - is it looping? it isn't done
+                        // - is it not looping? has more time elapsed than the computed time it should play for 
+                        //   based on sample count and frequency
+                        double playTime = (double)(timer_ns_gettime64() - channel->startTimeNs) * 1e-9;
                         double freq = (AUDIO_FREQUENCY / sfxList[channel->soundID].ratediv);
                         double samplesTime = (double)channel->sampleLength / freq;
-                        printf("xxx\t\tloop %d freq %f playTime %f samplesTime %f\n", channel->loop, freq, playTime, samplesTime);
+                        // this allows us to release channels ASAP and not rely solely on the eviction logic in PlaySfx
                         if ((!channel->loop) && (samplesTime <= playTime)) {
-                                printf("xxx\t\t\tReleased AICA Channel %d\n", aicaChannel);
                                 channel->state   = CHANNEL_IDLE;
                                 channel->soundID = -1;
                                 channel->aicaChannel = -1;
-                                channel->zeroPosCount = -1;
                                 snd_sfx_stop(aicaChannel);
                                 snd_sfx_chn_free(aicaChannel);
                         }
@@ -152,8 +111,7 @@ void AudioDeviceBase::ProcessAudioMixing(void *stream, int32 length)
             }
         }
     }
-#endif
-#if RETRO_PLATFORM != RETRO_KALLISTIOS
+#else
     SAMPLE_FORMAT *streamF    = (SAMPLE_FORMAT *)stream;
     SAMPLE_FORMAT *streamEndF = ((SAMPLE_FORMAT *)stream) + length;
 
@@ -267,10 +225,17 @@ void AudioDeviceBase::InitAudioChannels()
         channels[i].soundID = -1;
         channels[i].state   = CHANNEL_IDLE;
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
+        // reset to invalid hardware playback channel number
         channels[i].aicaChannel = -1;
-        channels[i].startNs = 0xFFFFFFFFFFFFFFFFL;
+        // reset to an invalid time
+        channels[i].startTimeNs = 0xFFFFFFFFFFFFFFFFL;
 #endif
     }
+
+#if RETRO_PLATFORM == RETRO_KALLISTIOS
+    // reserve the last channel for music stream
+    channels[CHANNEL_COUNT - 1].state = CHANNEL_STREAM;
+#endif
 
 #if RETRO_PLATFORM != RETRO_KALLISTIOS
     // Compute a lookup table of floating-point linear interpolation delta scales,
@@ -364,8 +329,8 @@ int32 RSDK::PlayStream(const char *filename, uint32 slot, uint32 startPos, uint3
     stream_destroy();
     stream_create(streamFilePath, loopPoint);
     stream_play();
-
-    return 0;
+    // last channel always reserved for music stream
+    return CHANNEL_COUNT - 1;
 #else
     if (!engine.streamsEnabled)
         return -1;
@@ -423,23 +388,6 @@ int32 RSDK::PlayStream(const char *filename, uint32 slot, uint32 startPos, uint3
 #define WAV_SIG_HEADER (0x46464952) // RIFF
 #define WAV_SIG_DATA   (0x61746164) // data
 
-#if 0
-//RETRO_PLATFORM == RETRO_KALLISTIOS
-struct snd_effect;
-LIST_HEAD(selist, snd_effect);
-
-typedef struct snd_effect {
-    uint32_t  locl, locr;
-    uint32_t  len;
-    uint32_t  rate;
-    uint32_t  used;
-    uint32_t  fmt;
-    uint16_t  stereo;
-
-    LIST_ENTRY(snd_effect)  list;
-} snd_effect_t;
-#endif
-
 void RSDK::LoadSfxToSlot(char *filename, uint8 slot, uint8 plays, uint8 scope)
 {
     char fullFilePath[0x80];
@@ -450,9 +398,12 @@ void RSDK::LoadSfxToSlot(char *filename, uint8 slot, uint8 plays, uint8 scope)
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
     sprintf_s(fullFilePath, sizeof(fullFilePath), "%s/Data/SoundFX/%s", KOS_USER_DIR, filename);
 
+    // sfxhnd_t is an opaque type
+    // but really is is a pointer to a `snd_effect_t` struct
+    // you have to poke at some KOS internals to get at it though
     sfxhnd_t hnd = snd_sfx_load(fullFilePath);
 
-    if (hnd != 0) {
+    if (hnd != (sfxhnd_t)0) {
         snd_effect_t *t = (snd_effect_t *)hnd;
         HASH_COPY_MD5(sfxList[slot].hash, hash);
         sfxList[slot].scope              = scope;
@@ -598,16 +549,13 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     if (sfx >= SFX_COUNT || !sfxList[sfx].scope)
         return -1;
 
-    if (sfx == -1)
-        return -1;
-
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
     if (sfxList[sfx].handle == (sfxhnd_t)0)
         return -1;
 
     int reservedChannel = snd_sfx_chn_alloc();
     if (reservedChannel == -1) {
-        printf("PlaySfx: reservedChannel == -1\n");
+        PrintLog(PRINT_ERROR, "PlaySfx: All AICA channels are reserved.\n");
         return -1;
     }
 
@@ -615,8 +563,8 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     // sound effect to play, by handle
     data.idx = sfxList[sfx].handle;
     data.chn = reservedChannel;
-    // 192 is chosen as maximum value to avoid clipping/distortion in sfx playback
-    data.vol = 192 * engine.soundFXVolume;
+    // 224 is chosen as maximum value to avoid clipping/distortion in sfx playback
+    data.vol = 224 * engine.soundFXVolume;
     // 127 is neutral/center pan
     data.pan = 127;
     // looping was causing problems
@@ -631,18 +579,14 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
             ++count;
     }
 
-    printf("*** playing sfx id: %d, count: %d\n", sfx, count);
-    
     int8 slot = -1;
     // if we've hit the max, replace the oldest one
     if (count >= sfxList[sfx].maxConcurrentPlays) {
-        printf("***\tmaxConcurrent reached\n");
         int32 highestStackID = 0;
-        for (int32 c = 0; c < CHANNEL_COUNT; ++c) {
+        for (int32 c = 0; c < CHANNEL_COUNT - 1; ++c) {
             if (channels[c].soundID == sfx) {
                 int32 stackID = sfxList[sfx].playCount - channels[c].playIndex;
-                if (stackID > highestStackID) {//} && channels[c].soundID == sfx) {
-                    printf("***\t\tdoing replacement %d -> %d\n", highestStackID, stackID);
+                if (stackID > highestStackID) {
                     slot           = c;
                     highestStackID = stackID;
                 }
@@ -652,18 +596,16 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
 
     int need_to_free = 0;
     if (slot >= 0) {
-//        snd_sfx_stop(channels[slot].aicaChannel);
+        // don't need the hardware channel we just tried to reserve
         snd_sfx_chn_free(reservedChannel);
-        printf("***\t\t\treplacement was going to go in aica ch # %d but now is playing on existing aica ch # %d\n", reservedChannel, channels[slot].aicaChannel);
+        // reuse the hardware channel from the oldest slot already playing this sfx
         data.chn = channels[slot].aicaChannel;
-//        return -1;
     }
 
     if (slot == -1) {
         // if we don't have a slot yet, try to pick any channel that's not currently playing
-        for (int32 c = 0; c < CHANNEL_COUNT; ++c) {
-            if (channels[c].soundID == -1) { //} && (channels[c].state == CHANNEL_SFX || channels[c].state == CHANNEL_IDLE)) { //!= CHANNEL_LOADING_STREAM) {
-                printf("***\t\tempty channel %d\n", c);
+        for (int32 c = 0; c < CHANNEL_COUNT - 1; ++c) {
+            if (channels[c].soundID == -1) {
                 slot = c;
                 break;
             }
@@ -673,37 +615,34 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     // as a last resort, run through all channels
     // pick the channel closest to being finished AND with lower priority
     if (slot < 0) {
-        printf("***\t\tas a last resort, run through all channels\n");
         uint32 len = 0xFFFFFFFF;
-        for (int32 c = 0; c < CHANNEL_COUNT; ++c) {
-                if (channels[c].sampleLength < len &&  priority > channels[c].priority /* && (channels[c].state == CHANNEL_SFX || channels[c].state == CHANNEL_IDLE) */) { //!= CHANNEL_LOADING_STREAM) {
-                printf("\t\t evicting channel %d for sfx %d\n", c, sfx);
+        for (int32 c = 0; c < CHANNEL_COUNT - 1; ++c) {
+            if (channels[c].sampleLength < len &&  priority > channels[c].priority) {
                 slot = c;
                 len  = (uint32)channels[c].sampleLength;
+                // we are going to evict the existing channel `c`
                 need_to_free = 1;
             }
         }
     }
 
     if (slot >= 0 && need_to_free) {
+        // evict whatever was in this slot
+        // release the hardware playback channel it was using
         if (channels[slot].aicaChannel != -1) {
             snd_sfx_stop(channels[slot].aicaChannel);
             snd_sfx_chn_free(channels[slot].aicaChannel);
-            printf("***\t\t\tReleased AICA Channel %d\n", channels[slot].aicaChannel);
             channels[slot].aicaChannel = -1;
-            channels[slot].zeroPosCount = -1;
         }
     }
 
     if (slot == -1) {
         snd_sfx_chn_free(reservedChannel);
-        printf("PlaySfx: slot == -1\n");
+        PrintLog(PRINT_ERROR, "PlaySfx: Could not find a usable ChannelInfo slot.\n");
         return -1;
     }
 
     channels[slot].state        = CHANNEL_SFX;
-    channels[slot].bufferPos    = 0;
-//    channels[slot].samplePtr    = sfxList[sfx].buffer;
     channels[slot].sampleLength = sfxList[sfx].length;
     channels[slot].volume       = 1.0f;
     channels[slot].pan          = 0.0f;
@@ -712,14 +651,16 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     if (loopPoint >= 2)
         channels[slot].loop = loopPoint;
     else
-        channels[slot].loop = 0; //loopPoint - 1;
+        channels[slot].loop = 0;
     channels[slot].priority  = priority;
     channels[slot].playIndex = sfxList[sfx].playCount++;
     channels[slot].aicaChannel = data.chn;
-    channels[slot].zeroPosCount = -1;
 
     snd_sfx_play_ex(&data);
-    channels[slot].startNs = timer_ns_gettime64();
+
+    // set the play time *after* firing off the AICA command
+    // so we don't end up cutting it off early
+    channels[slot].startTimeNs = timer_ns_gettime64();
 
     return slot;
 #else
@@ -789,36 +730,40 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
 
 void RSDK::SetChannelAttributes(uint8 channel, float volume, float panning, float speed)
 {
-//    return;
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
-    if ((channel < CHANNEL_COUNT) && (channels[channel].state == CHANNEL_SFX)) {
-        volume                   = fminf(4.0f, volume);
-        volume                   = fmaxf(0.0f, volume);
-        channels[channel].volume = volume;
+    volume                   = fminf(4.0f, volume);
+    volume                   = fmaxf(0.0f, volume);
+    channels[channel].volume = volume;
 
-        panning               = fminf(1.0f, panning);
-        panning               = fmaxf(-1.0f, panning);
-        channels[channel].pan = panning;
+    panning               = fminf(1.0f, panning);
+    panning               = fmaxf(-1.0f, panning);
+    channels[channel].pan = panning;
 
-        if (speed > 0.0f)
-            channels[channel].speed = (int32)(speed * TO_FIXED(1));
-        else if (speed == 1.0f)
-            channels[channel].speed = TO_FIXED(1);
+    if (speed > 0.0f)
+        channels[channel].speed = speed;
+    else if (speed == 1.0f)
+        channels[channel].speed = speed;
 
-        sfx_play_data_t data = {0};
-        // sound effect to play, by handle
-        data.idx = sfxList[channels[channel].soundID].handle;
-        data.chn = channels[channel].aicaChannel;
-        // 192 is chosen as maximum value to avoid clipping/distortion in sfx playback
-        data.vol = 192 * (channels[channel].volume /* * 0.25f */) * engine.soundFXVolume;
-        // 127 is neutral/center pan
-        data.pan = 127 * (channels[channel].pan + 1.0f);
-        // looping was causing problems
-        data.loop = channels[channel].loop != 0;
-        data.loopstart = channels[channel].loop;
-        data.freq = /* channels[channel]. */speed * (AUDIO_FREQUENCY / sfxList[channels[channel].soundID].ratediv);
-        snd_sfx_update_ex(&data);
-        printf("%%%%%% set channel %d to freq %d vol %d pan %d\n", channel, data.freq, data.vol, data.pan);
+    if ((channel < CHANNEL_COUNT - 1) && (channels[channel].state == CHANNEL_SFX)) {
+        // safety check, is there a hardware playback channel assigned to this ChannelInfo
+        if (channels[channel].aicaChannel != -1) {
+            sfx_play_data_t data = {0};
+            // sound effect to play, by handle
+            data.idx = sfxList[channels[channel].soundID].handle;
+            data.chn = channels[channel].aicaChannel;
+            // 224 is chosen as maximum value to avoid clipping/distortion in sfx playback
+            data.vol = 224 * (channels[channel].volume) * engine.soundFXVolume;
+            // 127 is neutral/center pan
+            data.pan = 127 * (channels[channel].pan + 1.0f);
+            // looping was causing problems
+            data.loop = channels[channel].loop != 0;
+            data.loopstart = channels[channel].loop;
+            data.freq = channels[channel].speed * (AUDIO_FREQUENCY / sfxList[channels[channel].soundID].ratediv);
+            snd_sfx_update_ex(&data);
+        }
+    } else if (channel == CHANNEL_COUNT - 1) {
+        // 224 is chosen as maximum value to avoid clipping/distortion in sfx playback
+        stream_volume(224 * channels[channel].volume * engine.streamVolume);
     }
 #endif
 #if RETRO_PLATFORM != RETRO_KALLISTIOS
@@ -842,16 +787,14 @@ void RSDK::SetChannelAttributes(uint8 channel, float volume, float panning, floa
 uint32 RSDK::GetChannelPos(uint32 channel)
 {
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
-    if (channel >= CHANNEL_COUNT)
+    if (channel >= CHANNEL_COUNT - 1)
         return 0;
 
     if (channels[channel].state == CHANNEL_SFX) {
         int aicaChannel = channels[channel].aicaChannel;
-        if (aicaChannel != -1) {
+        if (aicaChannel != -1)
             return snd_get_pos(aicaChannel);
-        }
     }
-
 #endif
 #if RETRO_PLATFORM != RETRO_KALLISTIOS
     if (channel >= CHANNEL_COUNT)
@@ -887,10 +830,14 @@ void RSDK::ClearStageSfx()
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
     stream_destroy();
 
-    for (int i = 0; i < 16; i++) {
+    for (int i = 0; i < CHANNEL_COUNT - 1; i++) {
         if (channels[i].aicaChannel != -1) {
             snd_sfx_stop(channels[i].aicaChannel);
             snd_sfx_chn_free(channels[i].aicaChannel);
+            // reset to invalid hardware playback channel number
+            channels[i].aicaChannel = -1;
+            // reset to an invalid time
+            channels[i].startTimeNs = 0xFFFFFFFFFFFFFFFFL;
         }
     }
 #endif
@@ -900,10 +847,6 @@ void RSDK::ClearStageSfx()
         if (channels[c].state == CHANNEL_SFX || channels[c].state == (CHANNEL_SFX | CHANNEL_PAUSED)) {
             channels[c].soundID = -1;
             channels[c].state   = CHANNEL_IDLE;
-#if RETRO_PLATFORM == RETRO_KALLISTIOS
-            channels[c].aicaChannel = -1;
-            channels[c].startNs = 0xFFFFFFFFFFFFFFFFL;
-#endif
         }
     }
 
@@ -927,10 +870,14 @@ void RSDK::ClearGlobalSfx()
 #if RETRO_PLATFORM == RETRO_KALLISTIOS
     stream_destroy();
 
-    for (int i = 0; i < 16; i++) {
+    for (int i = 0; i < CHANNEL_COUNT - 1; i++) {
         if (channels[i].aicaChannel != -1) {
             snd_sfx_stop(channels[i].aicaChannel);
             snd_sfx_chn_free(channels[i].aicaChannel);
+            // reset to invalid hardware playback channel number
+            channels[i].aicaChannel = -1;
+            // reset to an invalid time
+            channels[i].startTimeNs = 0xFFFFFFFFFFFFFFFFL;
         }
     }
 #endif
@@ -940,10 +887,6 @@ void RSDK::ClearGlobalSfx()
         if (channels[c].state == CHANNEL_SFX || channels[c].state == (CHANNEL_SFX | CHANNEL_PAUSED)) {
             channels[c].soundID = -1;
             channels[c].state   = CHANNEL_IDLE;
-#if RETRO_PLATFORM == RETRO_KALLISTIOS
-            channels[c].aicaChannel = -1;
-            channels[c].startNs = 0xFFFFFFFFFFFFFFFFL;
-#endif
         }
     }
