+++ RSDKv5/RSDK/Audio/KallistiOS/KallistiOSAudioDevice.cpp
@@ -1,5 +1,6 @@
 #include <dc/sound/sound.h>
 
+#include "KallistiOSSfxUpdate.cpp"
 #include "KallistiOSStream.cpp"
 
 static bool deviceInited;
